refactor: apply PR US-0 and US-0.1 feedback
